pt1_proyecto_sd/exporter/requirements.txt
pymongo==4.5.0
python-dotenv==1.0.0
